LayoutTests/fast/dom/Text/remove.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="test"></div>
<script>

description('This tests the DOM 4 remove method on a Text node.');

var testDiv = document.getElementById('test');
var text = document.createTextNode('text');
testDiv.appendChild(text);
shouldBe('testDiv.childNodes.length', '1');
text.remove();
shouldBe('testDiv.childNodes.length', '0');
text.remove();
shouldBe('testDiv.childNodes.length', '0');

</script>
<script src="../../js/resources/js-test-post.js"></script>
